Move and improve test
